fix: parsing empty array

+++ src/lexer.rs
@@ -21,7 +21,7 @@ pub struct Token {
     index: usize,
     pub lexeme: String,
     pub token_type: TokenType,
-    pub line: u8
+    pub line: u32
 }
 
 const NUMERICS: [char; 12] = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9', '.', '_'];
@@ -32,7 +32,7 @@ pub struct Lexer {
     source_content: Vec<char>,
     pub tokens: Vec<Token>,
     index: usize,
-    current_line: u8
+    current_line: u32
 } 
 
 impl Lexer {
